Refactor insertion handling logic to place all use of prepareContentForHydration behind prepareForInsertion which also also does toMapTree and validation
