+++ packages/dds/tree/src/feature-libraries/flex-tree/utilities.ts
@@ -11,8 +11,9 @@ import {
 	rootField,
 	type SchemaAndPolicy,
 } from "../../core/index.js";
+import type { FlexTreeContext } from "./context.js";
 
-import { TreeStatus, type FlexTreeEntity } from "./flexTreeTypes.js";
+import { TreeStatus } from "./flexTreeTypes.js";
 /**
  * Checks the detached field and returns the TreeStatus based on whether or not the detached field is a root field.
  * @param detachedField - the detached field you want to check.
@@ -80,10 +81,10 @@ export interface DetachedFieldCache {
  * @returns A {@link SchemaAndPolicy} object with the stored schema and policy from the node or field provided.
  * For {@link Unhydrated} nodes this schema may only describe the types allowed subtree for this particular entity.
  */
-export function getSchemaAndPolicy(nodeOrField: FlexTreeEntity): SchemaAndPolicy {
+export function getSchemaAndPolicy(context: FlexTreeContext): SchemaAndPolicy {
 	return {
-		schema: nodeOrField.context.schema,
-		policy: nodeOrField.context.schemaPolicy,
+		schema: context.schema,
+		policy: context.schemaPolicy,
 	};
 }
